refactor: 능력 버튼 리스트를 컴포넌트로 분리
